requirements.txt
Flask==2.2.3
Flask-WTF==1.1.1
PyMySQL==1.0.3
Jinja2==3.1.2
WTForms~=3.0.1
python-dotenv~=1.0.0
Werkzeug~=2.2.3